processor/upload/requirements.txt
minio
pymongo
pika
python-magic